third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-getBoxModel.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function test()
{
    var nodeInfo = {};
    InspectorTest.eventHandler["DOM.setChildNodes"] = setChildNodes;
    InspectorTest.sendCommand("DOM.enable", {});
    InspectorTest.sendCommand("DOM.getNodeForLocation", { "x": 100, "y": 200 }, onGetNodeForLocation);

    function onGetNodeForLocation(message)
    {
        if (message.error) {
            InspectorTest.log(message.error.message);
            InspectorTest.completeTest();
            return;
        }
        var ids = Object.keys(nodeInfo);
        for (var i = 0; i < ids.length; ++i)
            ids[i] = parseInt(ids[i]);
        ids.sort(function(a, b) { return a < b ? -1 : (a === b ? 0 : 1); });
        // Skip first - it does not have round values.
        for (var i = 1; i < ids.length; ++i)
            InspectorTest.sendCommand("DOM.getBoxModel", { "nodeId": ids[i] }, printBoxModel.bind(null, nodeInfo[ids[i]], i == ids.length - 1));
    }

    function printBoxModel(nodeInfo, lastNode, message)
    {
        if (message.error) {
            InspectorTest.log(nodeInfo.nodeName + ": " + message.error.message);
            if (lastNode)
                InspectorTest.completeTest();
            return;
        }
        InspectorTest.logObject(message.result.model.content, nodeInfo.nodeName + " " + nodeInfo.attributes + " ");
        if (lastNode)
            InspectorTest.completeTest();
    }

    function setChildNodes(message)
    {
        var nodes = message.params.nodes;
        for (var i = 0; i < nodes.length; ++i) {
            nodeInfo[nodes[i].nodeId] = nodes[i];
            delete nodes[i].nodeId;
        }
    }
}

</script>
</head>
<body onload="runTest()">
<div>
Several<br>
Lines<br>
Of<br>
Text<br>
<div style="position:absolute;top:100;left:0;width:100;height:100;background:red"></div>
<div style="position:absolute;top:200;left:100;width:100;height:100;background:green"></div>
<div style="position:absolute;top:150;left:50;width:100;height:100;background:blue;transform:rotate(45deg);"></div>
</div>
</body>
</html>
