third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-getBoxModel-expected.txt


#text: Could not compute box model.
BR: Could not compute box model.
#text: Could not compute box model.
BR: Could not compute box model.
#text: Could not compute box model.
BR: Could not compute box model.
#text: Could not compute box model.
BR: Could not compute box model.
DIV style,position:absolute;top:100;left:0;width:100;height:100;background:red [
    [0] : 0
    [1] : 100
    [2] : 100
    [3] : 100
    [4] : 100
    [5] : 200
    [6] : 0
    [7] : 200
]
DIV style,position:absolute;top:200;left:100;width:100;height:100;background:green [
    [0] : 100
    [1] : 200
    [2] : 200
    [3] : 200
    [4] : 200
    [5] : 300
    [6] : 100
    [7] : 300
]
DIV style,position:absolute;top:150;left:50;width:100;height:100;background:blue;transform:rotate(45deg); [
    [0] : 100
    [1] : 129
    [2] : 171
    [3] : 200
    [4] : 100
    [5] : 271
    [6] : 29
    [7] : 200
]
